Migrate tests into diagraph_tests
